Updated AutoSuggest Names
